fifo: allow large semaphore requests
Currently, `fifo.Semaphore.Acquire` errors out when a request is
larger than the capacity. This is not very useful, there is not much
the caller can do to handle this situation. For example, in Pebble we
want a semaphore that limits the number of outstanding bytes to read
but we don't want it to act as a hard limit on the block size.

This change adjusts the semaphore logic to allow acquisitions higher
than the capacity. They will acquire the entire capacity and hence
will run without other requests in parallel.

Note that the state of a semaphore having outstanding > capacity was
already possible when the capacity is changed dynamically.