cralloc: add ScratchBuffer
Add a helper type for reusing a buffer. This simplifies the typical
pattern of passing and returning a slice. It is also useful for cases
where a function might return either a slice from the scratch buffer
or some other aliased slice (for example, `EncodeMVCCValueForExport`
in cockroachdb has to return an extra bool so the caller can know if
it should update the scratch buffer).